// src/utils/enums/compute.enum.ts
export enum BattleLeagueCPType {
  Master,
  Little = 500,
  Great = 1500,
  Ultra = 2500,
  InsMaster = 10000,
}

export enum BattleLeagueIconType {
  Little = 'GBL_littlecup',
  Great = 'great_league',
  Ultra = 'ultra_league',
  Master = 'master_league',
}

export enum BattleLeagueTag {
  Seeker = 'seeker',
  SafariZone = 'safari_zone',
}

export enum FormType {
  Default,
  Shiny,
}